course_6/module_4_lab_2/requirements.txt
openai==0.28.1
sentence-transformers==2.3.1
pandas==2.2.0
uvicorn==0.27.1
qdrant-client==1.9.0
fastapi==0.109.2
load_dotenv